🎨 feat(code): Refactored imports and endpoints for improved modularity - Renamed `ai` module to `slack.ai` in multiple files. - Updated import paths accordingly in `chat.py`, `app.py`, and `ai.py`. - Added new endpoint `/slack/events` in `app.py` for processing Slack events.
